+++ elfo-core/src/dumping/config.rs
@@ -1,6 +1,6 @@
 use serde::Deserialize;
 
-#[derive(Clone, Deserialize)]
+#[derive(Debug, Clone, Deserialize)]
 #[serde(default)]
 pub(crate) struct DumpingConfig {
     pub(crate) disabled: bool,